+++ mutraction-dom-docs/src/nav.tsx
@@ -40,7 +40,6 @@ export const nav =
                                 <summary><a>DOM wrappers</a></summary>
                                 <ul>
                                     <li><a href="#ref/ForEach">ForEach()</a></li>
-                                    <li><a href="#ref/ForEachPersist">ForEachPersist()</a></li>
                                     <li><a href="#ref/PromiseLoader">PromiseLoader()</a></li>
                                     <li><a href="#ref/Swapper">Swapper()</a></li>
                                     <li><a href="#ref/Router">Router()</a></li>